added changes from the modified code to the other stuff in the program
see the discord for context ig

+++ mco1-sokobot/sokobot.bat
@@ -1,3 +1,10 @@
+@echo off
+if "%~1"=="" (
+    echo Usage: %~nx0 ^<map name^> [--gui]
+    exit /b 1
+)
 del /s /q *.class
 javac src/main/Driver.java -cp src
-java -classpath src main.Driver %1 bot
+set MODE=raw
+if "%2"=="--gui" set MODE=bot
+java -classpath src main.Driver %1 %MODE%
