+++ mco1-sokobot/sokobot.sh
@@ -1,4 +1,12 @@
 #!/bin/bash
+if [ -z "$1" ]; then
+    echo "Usage: $0 <map name> [--gui]"
+    exit 1
+fi
 find . -name "*.class" -type f -delete
 javac src/main/Driver.java -cp src
-java -classpath src main.Driver "$1" bot
+MODE="raw"
+if [ "$2" = "--gui" ]; then
+    MODE="bot"
+fi
+java -classpath src main.Driver "$1" $MODE
